rename rules module to strategies
