+++ sympy/strategies/branch/__init__.py
