+++ sympy/strategies/branch/core.py
