+++ sympy/strategies/branch/tools.py
