+++ sympy/strategies/tests/test_core.py
@@ -1,4 +1,4 @@
-from sympy.rules.core import (null_safe, exhaust, memoize, condition,
+from sympy.strategies.core import (null_safe, exhaust, memoize, condition,
         chain, tryit, do_one, debug, switch, minimize)
 from functools import partial
