+++ sympy/strategies/tools.py
@@ -8,7 +8,7 @@ def subs(d):
     Example
     =======
 
-    >>> from sympy.rules.tools import subs
+    >>> from sympy.strategies.tools import subs
     >>> from sympy import Basic
     >>> mapping = {1: 4, 4: 1, Basic(5): Basic(6, 7)}
     >>> expr = Basic(1, Basic(2, 3), Basic(4, Basic(5)))
@@ -35,7 +35,7 @@ def typed(ruletypes):
     inputs:
         ruletypes -- a dict mapping {Type: rule}
 
-    >>> from sympy.rules import rm_id, typed
+    >>> from sympy.strategies import rm_id, typed
     >>> from sympy import Add, Mul
     >>> rm_zeros = rm_id(lambda x: x==0)
     >>> rm_ones  = rm_id(lambda x: x==1)
