+++ sympy/strategies/util.py
